+++ docs/source/history.md
@@ -414,6 +414,7 @@ Enjoy - we are awaiting comments and issues, thanks for testing.
 - Wake models:
   - Removing `sbeta` limitation from `Bastankhah2014` and `CrespoHernandez`
   - New option `regularize` for `Bastankhah2014`, `TurbOParkWake` and `TurbOParkIXWake`, smoothing the sqrt at the end of the near wake region
+  - Introducing `WakeMirror` wrapper around wake models, modelling wake reflection from ground or horizontal plane via virtual mirrored turbines
 - Tests:
   - Fresh `flappy` v0.6.2 test data for all Bastankhah and CrespoHernandez wakes, also without the `sbeta` limitation